+++ CHANGELOG.md
@@ -7,13 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * Outlookbar (NavigationBar, Accordion)
 * Pageinator (PagingControl)
 * DynamicTabControl
-* HeaderItemsControl
 
 ## [Unreleased]
 * Added ArcPanel which arranges child elements in a configurable arc form.
 * Added EllipsePanel which arranges child elements in a ellipse form.
 * Added UniformWrapPanel which enhances the WrapPanel by the feature that all items will have the same size.
 * Added Resizer which brings the possibility to resize every UI control manually by hold and drag the corners or sides.
+* Added HeaderItemsControl to allign items with a header and content vertically or horizontally.
 
 ## [1.0.0] - 2021-06-13
 ### Added